:recycle: Refactor: Vision Utils
